// frontend/src/generated/ToplineSource.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ToplineLine } from "./ToplineLine";
import type { VoicePreset } from "./VoicePreset";

export type ToplineSource = { section_name: string, voice: VoicePreset, lines: Array<ToplineLine>, };